progress
tiny fixes and code added, may be changed later as bits get closer to
completion

+++ src/core/Pattern.java
@@ -12,6 +12,7 @@ import util.jython.*;
  */
 public class Pattern implements Destructable,TransientContainer<Composition> {
 	private static final long serialVersionUID = 1L;
+	
 	/**
 	 * Basically, how many steps in a measure
 	 */
@@ -32,23 +33,30 @@ public class Pattern implements Destructable,TransientContainer<Composition> {
 	public final ArrayList<int[]> clips;
 	
 	/** 
-	 * Byte encoding:
+	 * Int encoding:
 	 * <br>
-	 * -1  - built in (Java)
+	 * 0 is unset
 	 * <br>
-	 * -2  - included (Python)
+	 * Negative is system managed:
+	 * <ul>
+	 * <li>-1 is built in (Java)</li>
+	 * <li>-2 is included (Python)</li>
+	 * <li>-3 is project specific (Python) with file path relative to project file</li>
+	 * </ul>
 	 * <br>
-	 * -3  - project specific (Python)
-	 * <br>
-	 * 64+ - reserved for user defined libraries (Python)
+	 * Positive is user managed:
+	 * <ul>
+	 * <li>1-999999 is reserved for user defined libraries (Python),
+	 * the number determines where to look with file path relative to executable</li>
+	 * </ul>
 	 */
-	protected byte synthType;
+	protected int synthType;
 	/**
 	 * The synthesizer to be used for everything
 	 * <br>
 	 * This may be shared
 	 * <br>
-	 * Based on byte code:
+	 * Based on int code:
 	 * if built in, string is the name, otherwise it's a filename
 	 */
 	protected String synthName;
